refactor: move the data loading logic and pagination to a separate files
